photo_manipulation_streamlit/pyproject.toml
[project]
name = "photo-manipulation-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.2.1",
    "streamlit>=1.45.0",
]
